requirements.txt
accelerate==0.34.2
aioboto3==13.1.1
aiobotocore==2.13.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioitertools==0.12.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
bitsandbytes==0.43.3
boto3==1.34.131
botocore==1.34.131
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cohere==5.9.2
contourpy==1.3.0
cycler==0.12.1
datasets==3.0.0
deeplake==3.9.23
deepspeed==0.15.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
fastavro==1.9.7
filelock==3.16.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.24.7
humbug==0.3.2
idna==3.8
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
libdeeplake==0.0.145
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
neural_compressor==3.0.2
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
onnx==1.16.2
openai==1.45.0
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
pathos==0.3.2
peft==0.12.0
pillow==10.2.0
platformdirs==4.3.2
pox==0.3.4
ppft==1.7.6.8
prettytable==3.11.0
protobuf==5.28.1
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pycocotools==2.0.8
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.8.1
s3transfer==0.10.2
safetensors==0.4.5
schema==0.7.7
scikit-learn==1.5.2
scipy==1.14.1
sentry-sdk==2.14.0
setproctitle==1.3.3
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sympy==1.13.2
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.4.1
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
trl==0.10.1
types-requests==2.32.0.20240914
typing_extensions==4.12.2
tyro==0.8.10
tzdata==2024.1
urllib3==2.2.3
wandb==0.18.0
wcwidth==0.2.13
wrapt==1.16.0
xxhash==3.5.0
yarl==1.11.1
